LPS-108987 Make client treat panels as a map, not a list
Seeing as that is what the server is giving to us. Note: this is a
divergence from the way it is done in layout-content-page-editor, where
we get a list, but for now it is easier for me to unbreak the client
with this quick fix than it is for me to try unifying things on the
server.